requirements.txt
streamlit
torch
torchvision
Pillow
flask
Werkzeug